vizro-core/examples/scratch_dev/requirements.txt
dash>=2.6.0
dash-bootstrap-components>=1.2.1
dash-mantine-components>=0.12.1
dash-iconify>=0.1.2
markdownify>=0.11.6
python-dotenv>=0.21.0
openai>=1.3.0
marked>=4.3.0  # For JavaScript markdown parsing
